// planning-idsdp-pps-frontend/src/app/main/modules/dpp-tapp-management/models/mtbf-fiscal-year-details-list-model.model.ts
export class MtbfFiscalYearDetailsListModel {
    uuid: string;
    fiscalYear: string;
    resourcesInFiveYearPlan: string;
    allocationAndProjection: string;
    allocationRequiredForProject: string;
    allocationOfApprovedProject: string;
    allocationOfRecommendedProject: string;
    allocationRequiredForProposedProject: string;
    totalRequirementOfAllocation: string;
    fiscalSpace: string;
}